+++ .github/workflows/ci.yml
@@ -26,7 +26,7 @@ jobs:
         uses: actions/setup-node@v4
         with:
           node-version: 22
-          cache: 'npm'
+          cache: "npm"
       - name: Install dependencies
         run: npm ci
       - name: Check for unformatted files
@@ -41,7 +41,7 @@ jobs:
         uses: actions/setup-node@v4
         with:
           node-version: 22
-          cache: 'npm'
+          cache: "npm"
       - name: Install dependencies
         run: npm ci
       - name: Lint JavaScript
@@ -60,7 +60,7 @@ jobs:
         uses: actions/setup-node@v4
         with:
           node-version: 22
-          cache: 'npm'
+          cache: "npm"
       - name: Install dependencies
         run: npm ci
       - name: Build
@@ -77,7 +77,7 @@ jobs:
         uses: actions/setup-node@v4
         with:
           node-version: 22
-          cache: 'npm'
+          cache: "npm"
       - name: Install dependencies
         run: npm ci
       - name: Build project
@@ -94,7 +94,7 @@ jobs:
         uses: actions/setup-node@v4
         with:
           node-version: 22
-          cache: 'npm'
+          cache: "npm"
       - name: Install dependencies
         run: npm ci
       - name: Build
@@ -112,7 +112,7 @@ jobs:
         uses: actions/setup-node@v4
         with:
           node-version: 22
-          cache: 'npm'
+          cache: "npm"
       - name: Install dependencies
         run: npm ci
       - name: Build storybook
@@ -129,7 +129,7 @@ jobs:
       - name: Run VRT
         uses: docker://mcr.microsoft.com/playwright:v1.47.2-jammy
         env:
-          STORYBOOK_URL: 'http://172.17.0.1:6006'
+          STORYBOOK_URL: "http://172.17.0.1:6006"
         with:
           args: npx playwright test --grep @vrt --shard="${{ matrix.shard }}/${{ strategy.job-total }}"
       - name: Stop storybook
@@ -152,7 +152,7 @@ jobs:
         uses: actions/setup-node@v4
         with:
           node-version: 22
-          cache: 'npm'
+          cache: "npm"
       - name: install dependencies
         run: npm ci
       - name: download all reports
@@ -188,7 +188,7 @@ jobs:
         uses: actions/setup-node@v4
         with:
           node-version: 22
-          cache: 'npm'
+          cache: "npm"
       - name: Install dependencies
         run: npm ci
       - name: Build storybook
@@ -205,7 +205,7 @@ jobs:
       - name: Run VRT
         uses: docker://mcr.microsoft.com/playwright:v1.47.2-jammy
         env:
-          STORYBOOK_URL: 'http://172.17.0.1:6006'
+          STORYBOOK_URL: "http://172.17.0.1:6006"
         with:
           args: npx playwright test --grep @vrt --shard="${{ matrix.shard }}/${{ strategy.job-total }}"
       - name: Stop storybook
@@ -228,7 +228,7 @@ jobs:
         uses: actions/setup-node@v4
         with:
           node-version: 22
-          cache: 'npm'
+          cache: "npm"
       - name: install dependencies
         run: npm ci
       - name: download all reports
@@ -248,19 +248,39 @@ jobs:
         if: ${{ needs.vrt-runner-all-flags.result == 'failure' }}
         run: exit 1
 
+  build-components-json:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+      - name: Use Node.js 20.x
+        uses: actions/setup-node@v4
+        with:
+          node-version: 22
+          cache: "npm"
+      - name: Install dependencies
+        run: npm ci
+      - name: Build storybook to generate story IDs
+        run: npm run build:storybook
+        working-directory: packages/react
+      - name: Build components.json
+        run: npx tsx script/components-json/build.ts --storybook-data 'storybook-static/index.json'
+        working-directory: packages/react
+
   aat-runner:
     runs-on: ubuntu-latest-16-cores
     strategy:
       fail-fast: false
       matrix:
         shard: [1, 2, 3, 4]
+    needs: build-components-json
     steps:
       - uses: actions/checkout@v4
       - name: Use Node.js 20.x
         uses: actions/setup-node@v4
         with:
           node-version: 22
-          cache: 'npm'
+          cache: "npm"
       - name: Install dependencies
         run: npm ci
       - name: Build storybook
@@ -277,7 +297,7 @@ jobs:
       - name: Run AAT
         uses: docker://mcr.microsoft.com/playwright:v1.47.2-jammy
         env:
-          STORYBOOK_URL: 'http://172.17.0.1:6006'
+          STORYBOOK_URL: "http://172.17.0.1:6006"
         with:
           args: npx playwright test --grep @aat --shard="${{ matrix.shard }}/${{ strategy.job-total }}"
       - name: Stop storybook
@@ -300,7 +320,7 @@ jobs:
         uses: actions/setup-node@v4
         with:
           node-version: 22
-          cache: 'npm'
+          cache: "npm"
       - name: install dependencies
         run: npm ci
       - name: download all reports
@@ -336,7 +356,7 @@ jobs:
         uses: actions/setup-node@v4
         with:
           node-version: 22
-          cache: 'npm'
+          cache: "npm"
       - name: Install dependencies
         run: npm ci
       - name: Build storybook
@@ -353,7 +373,7 @@ jobs:
       - name: Run AAT
         uses: docker://mcr.microsoft.com/playwright:v1.47.2-jammy
         env:
-          STORYBOOK_URL: 'http://172.17.0.1:6006'
+          STORYBOOK_URL: "http://172.17.0.1:6006"
         with:
           args: npx playwright test --grep @aat --shard="${{ matrix.shard }}/${{ strategy.job-total }}"
       - name: Stop storybook
@@ -376,7 +396,7 @@ jobs:
         uses: actions/setup-node@v4
         with:
           node-version: 22
-          cache: 'npm'
+          cache: "npm"
       - name: install dependencies
         run: npm ci
       - name: download all reports
@@ -396,25 +416,6 @@ jobs:
         if: ${{ needs.aat-runner-all-flags.result == 'failure' }}
         run: exit 1
 
-  build-components-json:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-      - name: Use Node.js 20.x
-        uses: actions/setup-node@v4
-        with:
-          node-version: 22
-          cache: 'npm'
-      - name: Install dependencies
-        run: npm ci
-      - name: Build storybook to generate story IDs
-        run: npm run build:storybook
-        working-directory: packages/react
-      - name: Build components.json
-        run: npx tsx script/components-json/build.ts --storybook-data 'storybook-static/index.json'
-        working-directory: packages/react
-
   sizes:
     runs-on: ubuntu-latest
     steps:
@@ -424,7 +425,7 @@ jobs:
         uses: actions/setup-node@v4
         with:
           node-version: 22
-          cache: 'npm'
+          cache: "npm"
       - name: Install dependencies
         run: npm ci
       - name: Build
