+++ packages/react/src/__tests__/storybook.test.tsx
@@ -1,8 +1,86 @@
 import glob from 'fast-glob'
-import {stories, ROOT_DIRECTORY, allowlist} from '../utils/testing'
 import groupBy from 'lodash.groupby'
 import fs from 'node:fs'
 import path from 'node:path'
+
+const ROOT_DIRECTORY = path.resolve(__dirname, '..', '..')
+
+// Components opted into the new story format
+// TODO: Remove this allowlist when all components use the new story format
+const allowlist = [
+  'ActionList',
+  'ActionMenu',
+  'AnchoredOverlay',
+  'Autocomplete',
+  'Avatar',
+  'AvatarStack',
+  'AvatarPair',
+  'Breadcrumbs',
+  'BranchName',
+  'Blankslate',
+  'Box',
+  'Button',
+  'Checkbox',
+  'CheckboxGroup',
+  'ConfirmationDialog',
+  'CounterLabel',
+  'DataTable',
+  'Details',
+  'Dialog',
+  'Flash',
+  'FormControl',
+  'Header',
+  'Heading',
+  'IconButton',
+  'FilteredActionList',
+  'Link',
+  'Octicon',
+  'Pagehead',
+  'Pagination',
+  'ProgressBar',
+  'Radio',
+  'RadioGroup',
+  'RelativeTime',
+  'Select',
+  'SegmentedControl',
+  'Spinner',
+  'StateLabel',
+  'SubNav',
+  'TabNav',
+  'Textarea',
+  'TextInput',
+  'TextInputWithTokens',
+  'Tooltip',
+  'TreeView',
+  'Timeline',
+  'ToggleSwitch',
+  'Token',
+  'UnderlineNav2',
+]
+
+const stories = glob
+  .sync('src/**/*.stories.tsx', {
+    cwd: ROOT_DIRECTORY,
+  })
+  // Filter out deprecated stories
+  .filter(file => !file.includes('deprecated'))
+  .filter(file =>
+    allowlist.some(
+      component => file.includes(`/${component}.stories.tsx`) || file.includes(`/${component}.features.stories.tsx`),
+    ),
+  )
+  .map(file => {
+    const filepath = path.join(ROOT_DIRECTORY, file)
+    const type = path.basename(filepath, '.stories.tsx').endsWith('features') ? 'feature' : 'default'
+    const name = type === 'feature' ? path.basename(file, '.features.stories.tsx') : path.basename(file, '.stories.tsx')
+
+    return {
+      name,
+      story: require(filepath),
+      type,
+      relativeFilepath: path.relative(ROOT_DIRECTORY, filepath),
+    }
+  })
 
 const components = Object.entries(
   groupBy(stories, ({name}) => {
