+++ packages/react/src/utils/testing.tsx
@@ -8,8 +8,6 @@ import customRules from '@github/axe-github'
 import {ThemeProvider} from '..'
 import {default as defaultTheme} from '../theme'
 import type {LiveRegionElement} from '@primer/live-region-element'
-import path from 'path'
-import {glob} from 'fast-glob'
 
 type ComputedStyles = Record<string, string | Record<string, string>>
 
@@ -17,80 +15,6 @@ type ComputedStyles = Record<string, string | Record<string, string>>
 const readFile = promisify(require('fs').readFile)
 
 export const COMPONENT_DISPLAY_NAME_REGEX = /^[A-Z][A-Za-z]+(\.[A-Z][A-Za-z]+)*$/
-
-export const ROOT_DIRECTORY = path.resolve(__dirname, '..', '..')
-
-// Components opted into the new story format
-// TODO: Remove this allowlist when all components use the new story format
-export const allowlist = [
-  'ActionList',
-  'ActionMenu',
-  'AnchoredOverlay',
-  'Autocomplete',
-  'Avatar',
-  'AvatarStack',
-  'AvatarPair',
-  'Breadcrumbs',
-  'BranchName',
-  'Blankslate',
-  'Box',
-  'Button',
-  'Checkbox',
-  'CheckboxGroup',
-  'ConfirmationDialog',
-  'CounterLabel',
-  'DataTable',
-  'Details',
-  'Dialog',
-  'Flash',
-  'FormControl',
-  'Header',
-  'Heading',
-  'IconButton',
-  'FilteredActionList',
-  'Link',
-  'Octicon',
-  'Pagehead',
-  'Pagination',
-  'ProgressBar',
-  'Radio',
-  'RadioGroup',
-  'RelativeTime',
-  'Select',
-  'SegmentedControl',
-  'Spinner',
-  'StateLabel',
-  'SubNav',
-  'TabNav',
-  'Textarea',
-  'TextInput',
-  'TextInputWithTokens',
-  'Tooltip',
-  'TreeView',
-  'Timeline',
-  'ToggleSwitch',
-  'Token',
-  'UnderlineNav2',
-]
-
-export const stories = glob
-  .sync('src/**/*.stories.tsx', {
-    cwd: ROOT_DIRECTORY,
-  })
-  // Filter out deprecated stories
-  .filter(file => !file.includes('deprecated'))
-  .filter(file =>
-    allowlist.some(
-      component => file.includes(`/${component}.stories.tsx`) || file.includes(`/${component}.features.stories.tsx`),
-    ),
-  )
-  .map(file => {
-    const filepath = path.join(ROOT_DIRECTORY, file)
-    const type = path.basename(filepath, '.stories.tsx').endsWith('features') ? 'feature' : 'default'
-    const name = type === 'feature' ? path.basename(file, '.features.stories.tsx') : path.basename(file, '.stories.tsx')
-
-    return {name, story: require(filepath), type, relativeFilepath: path.relative(ROOT_DIRECTORY, filepath)}
-  })
 
 declare global {
   // eslint-disable-next-line @typescript-eslint/no-namespace
